+++ packages/resources/src/NextjsSite.ts
@@ -573,7 +573,16 @@ export class NextjsSite extends Construct implements SSTConstruct {
     });
 
     // Create alias
-    fn.currentVersion.addAlias("live");
+    // note: Alias used to be created using on the Version using
+    //       "fn.currentVersion.addAlias()". It was deprecated.
+    //       To preserve the logical id, we are going to manually create
+    //       the Alias on the version, and use "Aliaslive" as id to preserve
+    //       the logical id.
+    const version = fn.currentVersion;
+    new lambda.Alias(version, "Aliaslive", {
+      aliasName: "live",
+      version,
+    });
 
     // Deploy after the code is updated
     if (hasRealCode) {
@@ -1212,8 +1221,9 @@ export class NextjsSite extends Construct implements SSTConstruct {
     }
 
     // Create custom resource
-    const waitForInvalidation =
-      this.props.waitForInvalidation === false ? false : true;
+    const waitForInvalidation = this.isPlaceholder
+      ? false
+      : (this.props.waitForInvalidation === false ? false : true);
     return new CustomResource(this, "CloudFrontInvalidation", {
       serviceToken: invalidator.functionArn,
       resourceType: "Custom::SSTCloudFrontInvalidation",