+++ packages/resources/src/StaticSite.ts
@@ -779,8 +779,9 @@ export class StaticSite extends Construct implements SSTConstruct {
       .digest("hex");
 
     // Create custom resource
-    const waitForInvalidation =
-      this.props.waitForInvalidation === false ? false : true;
+    const waitForInvalidation = this.isPlaceholder
+      ? false
+      : (this.props.waitForInvalidation === false ? false : true);
     return new CustomResource(this, "CloudFrontInvalidation", {
       serviceToken: invalidator.functionArn,
       resourceType: "Custom::SSTCloudFrontInvalidation",