+++ packages/cli2/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @serverless-stack/cli2
 
+## 1.15.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @serverless-stack/node@1.15.16
+
+## 1.15.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @serverless-stack/node@1.15.15
+
+## 1.15.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @serverless-stack/node@1.15.14
+
 ## 1.15.13
 
 ### Patch Changes